LayoutTests/fast/viewport/viewport-40-expected.txt

CONSOLE MESSAGE: line 3: Viewport width or height set to physical device width, try using "device-width" constant instead for future compatibility.
viewport size 700x480 scale 0.733333 with limits [0.733333, 5.000000]
